Check Point System Package with Event Handler
Two Checkpoints already with tags and the dynamic checkpoint system in an event handler. This event handler will be the thing that handles the animation plays later on for now it's just a placeholder for interactions. Their is a game object with the checkpoint system in place all you would need to do is simply import the package and then do small minor adjustments. Word of Advice the Player is not tagged as player so it may not work until this is changed.