I added a Respawn
It checks to see if you triggered a collision with something with tag "Death" and if it does to respawn you. Right now only the Falling respawn works I don't know how to add objects to the scene they keep disappearing when I play. If someone can verify gameobjects with colliders and the tag's "Death" actually respawn the player I would greatly appreciate it.