Respawn Changes
Should kill the player after 3000 delta time units. I think this is also very low so increasing the number to be something based on delta time is crucial to having it play the animation. I need some assistance in adding the PostFX Layer and Volume to the player "caught" animation.